First pass at trying to come up with a parser.
Clearly the tests are not tests, but they are good
building blocks to making things happen.

The general principle is being able to parse some input
and then generate output from those tokens. Most of the
rules used here are borrowed from tiddlywebplugins.links.

Some issues are apparent:

* If the linked output is not the same length as ones
  delicately crafted line length in the input that is
  annoying, that suggest that links which are anything
  other than wikilinks are to be spurned. This is probably
  the case anyway, so maybe not a big deal.

* I forget the other one.

The main code of interest is in test/test_tdd.py. Which is
not a test. The generation code in there is not how it will
be in the finaly form, but the concept is about right.

Because whitespace is significant in the parser we don't
worry about block v phrase level parsing. It's just tokens.

So that brings us around to why use a parser in the first
place? Well, because it makes changing the syntax rules
easy, without having any impact on the output rules. And
that's just part of it.